services/scoreboard/requirements.txt
strawberry-graphql>=0.180.5
click>=8.1.3
